requirements.txt
Flask>=3.0,<4
Flask-Limiter>=3.5,<4
python-dotenv>=1.0,<2
langchain>=0.2,<0.3
langchain-community>=0.2,<0.3
langchain-openai>=0.1.7,<0.2
openai>=1.40,<2
chromadb>=0.5,<0.6
python-docx>=1.1,<1.2
tiktoken>=0.7,<0.8
colorama>=0.4,<0.5
gunicorn>=21.2,<22
pypandoc>=1.12,<2
# BM25 retriever dependency used by langchain_community
rank-bm25>=0.2,<0.3
# Optional: Redis client for limiter/task state
redis>=5,<6
# RQ for multi-worker background jobs
rq>=1.16,<2
# chunking-evaluation  # ensure your internal/private package is available or vendor its code
